Fixed up pin init stuff, still no measurements.

+++ Exercise10p3-US_Sensor/timers.h
@@ -5,7 +5,7 @@ void pwm_init(TIM_TypeDef * TIMx, uint32_t timerperiph, uint32_t prescaler_div,
     uint32_t period,uint16_t countermode, int channel);
 
 void init_input_pw_capture(TIM_TypeDef * TIMx, uint32_t timerperiph, uint32_t prescaler_div,
-    uint32_t period,uint16_t countermode, int channel,uint16_t polarity, uint16_t selection,
+    uint32_t period,uint16_t countermode, uint16_t channel, uint16_t polarity, uint16_t selection,
     uint16_t input_trigger, bool init_complete);
 
 #endif
